fix: GoalのHealthの型を修正

+++ Assets/akanevrc/TowerDefence/Runtime/Models/Goal/Goal.cs
@@ -5,6 +5,6 @@ namespace akanevrc.TowerDefence
     [Serializable]
     public struct Goal
     {
-        public int Health;
+        public float Health;
     }
 }
